New project structure and new Main to allocate memory effectively
