--- src/main/java/fr/su/handlers/insertion/RemoteInsertionHandler.java
@@ -1,22 +1,0 @@
-package fr.su.handlers.insertion;
-
-import fr.su.handlers.ForwardingManager;
-import jakarta.inject.Inject;
-import jakarta.inject.Singleton;
-import jakarta.ws.rs.core.Response;
-
-import java.io.File;
-import java.io.IOException;
-
-@Singleton
-public class RemoteInsertionHandler implements InsertionHandler {
-
-    @Inject
-    ForwardingManager forwardingManager;
-
-    @Override
-    public int insert(File file) throws IOException {
-        Response response = forwardingManager.forwardInsert(file);
-        return response == null ? 200 : response.getStatus();
-    }
-}
