--- src/main/java/fr/su/handlers/select/tree/Operator.java
@@ -1,6 +1,0 @@
-package fr.su.handlers.select.tree;
-
-public enum Operator {
-
-    AND, OR, EQUALS
-}
