--- src/main/java/fr/su/handlers/table/TableHandler.java
@@ -1,12 +1,0 @@
-package fr.su.handlers.table;
-
-import fr.su.controllers.TableController;
-import jakarta.ws.rs.core.Response;
-
-import java.io.IOException;
-
-public interface TableHandler {
-
-    public TableController.TableBody createTable(TableController.TableBody tableBody) throws IOException;
-
-}
