+++ src/main/java/fr/su/memorydb/Main.java
@@ -1,0 +1,13 @@
+package fr.su.memorydb;
+
+import io.quarkus.runtime.Quarkus;
+import io.quarkus.runtime.annotations.QuarkusMain;
+
+@QuarkusMain
+public class Main {
+
+    public static void main(String[] args) {
+        Quarkus.run(args);
+    }
+
+}
