+++ src/main/java/fr/su/memorydb/handlers/insertion/LambdaInsertion.java
@@ -1,0 +1,10 @@
+package fr.su.memorydb.handlers.insertion;
+
+import org.apache.parquet.example.data.Group;
+
+@FunctionalInterface
+public interface LambdaInsertion {
+
+    Object call(Group g, String field, int index);
+
+}
