+++ src/main/java/fr/su/memorydb/handlers/select/tree/Operator.java
@@ -1,0 +1,6 @@
+package fr.su.memorydb.handlers.select.tree;
+
+public enum Operator {
+
+    AND, OR, EQUALS
+}
