+++ src/main/java/fr/su/memorydb/utils/exceptions/NameAlreadyUsedException.java
@@ -1,0 +1,4 @@
+package fr.su.memorydb.utils.exceptions;
+
+public class NameAlreadyUsedException extends Exception {
+}
