+++ src/main/java/fr/su/memorydb/utils/exceptions/TableColumnSizeException.java
@@ -1,0 +1,8 @@
+package fr.su.memorydb.utils.exceptions;
+
+public class TableColumnSizeException extends Exception {
+
+    public TableColumnSizeException(int dbNumber, int fileNumber) {
+        super("Number of columns is not the same between our database (" + dbNumber + ") and your file (" + fileNumber + ")");
+    }
+}
