+++ src/main/java/fr/su/memorydb/utils/exceptions/UnknownColumnTypeException.java
@@ -1,0 +1,4 @@
+package fr.su.memorydb.utils.exceptions;
+
+public class UnknownColumnTypeException extends Exception {
+}
