+++ src/main/java/fr/su/memorydb/utils/exceptions/WrongTableFormatException.java
@@ -1,0 +1,7 @@
+package fr.su.memorydb.utils.exceptions;
+
+public class WrongTableFormatException extends Exception {
+    public WrongTableFormatException() {
+        super("Parquet Shema is not same as DataBase Shema");
+    }
+}
