--- src/main/java/fr/su/utils/exceptions/NameAlreadyUsedException.java
@@ -1,4 +1,0 @@
-package fr.su.utils.exceptions;
-
-public class NameAlreadyUsedException extends Exception {
-}
