--- src/main/java/fr/su/utils/exceptions/TableColumnSizeException.java
@@ -1,8 +1,0 @@
-package fr.su.utils.exceptions;
-
-public class TableColumnSizeException extends Exception {
-
-    public TableColumnSizeException(int dbNumber, int fileNumber) {
-        super("Number of columns is not the same between our database (" + dbNumber + ") and your file (" + fileNumber + ")");
-    }
-}
