--- src/main/java/fr/su/utils/exceptions/UnknownColumnTypeException.java
@@ -1,4 +1,0 @@
-package fr.su.utils.exceptions;
-
-public class UnknownColumnTypeException extends Exception {
-}
