--- src/main/java/fr/su/utils/exceptions/WrongTableFormatException.java
@@ -1,7 +1,0 @@
-package fr.su.utils.exceptions;
-
-public class WrongTableFormatException extends Exception {
-    public WrongTableFormatException() {
-        super("Parquet Shema is not same as DataBase Shema");
-    }
-}
